Moved test to RegressionTests
